Promisify: Moved methods to promisify.js
